refactored into middleware
