+++ functions/index.js
@@ -1,9 +1,7 @@
 const functions = require("firebase-functions");
-const axios = require("axios");
+const { getGeocodes } = require("./middleware");
 
 exports.getGeo = functions.https.onRequest(async (request, response) => {
-  const locationEndpoint = functions.config().location.endpoint;
-  const apikey = functions.config().location.key;
   const address = request.query.address;
 
   if (address === undefined) {
@@ -13,37 +11,28 @@ exports.getGeo = functions.https.onRequest(async (request, response) => {
       message: "Params Error: missing or incorrect address format",
     });
   }
-
-  return await axios
-    .get(locationEndpoint, {
-      params: {
-        address,
-        key: apikey,
-      },
-    })
-    .then((resp) => {
-      if (resp.data.status === "REQUEST_DENIED") {
-        return response.status(500).send({
-          error: true,
-          status: 500,
-          message: "Invalid API Key Set!",
-        });
-      }
-
+  return await getGeocodes({ functions, address })
+    .then((res) => {
       return response.status(200).send({
         error: false,
         status: 200,
-        message: {
-          geo: resp.data.results[0].geometry.location,
-          id: resp.data.results[0].place_id,
-        },
+        message: res.message,
       });
     })
     .catch((err) => {
-      return response.status(500).send({
-        error: true,
-        status: 500,
-        message: `Internal Server Error: ${err.response.data.error_message}`,
-      });
+      const message = err.message;
+      if (message.includes("REQUEST_DENIED")) {
+        return response.status(500).send({
+          error: true,
+          status: 500,
+          message: message,
+        });
+      } else {
+        return response.status(400).send({
+          error: true,
+          status: 400,
+          message: message,
+        });
+      }
     });
 });
